+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+## [2026-03-27] - Release
+- Triggered release pipeline for lite, diagramly, full (macro title → "Diagram (Mermaid, PlantUML & ZenUML)")
+
+## [2026-03-27] - Release
+- Triggered release pipeline for full-forge
+
 ## [2026-03-26] - Release
 - Triggered release pipeline for lite
+
+## [2026-04-05] - Release
+- Triggered release pipeline for diagramly
+
+## [2026-04-05] - Release
+- Triggered release pipeline for full